fix(ExchangeRateService): change int to long in doesRateHaveMistake()

+++ src/main/java/service/ExchangeRateService.java
@@ -253,15 +253,9 @@ public class ExchangeRateService {
 
     private boolean doesRateHaveMistake(BigDecimal rate) {
         BigDecimal normalizedRate = rate.stripTrailingZeros();
-        int integerDigits = Math.max(
-                normalizedRate.precision() - normalizedRate.scale(),
-                0
-        );
+        long integerDigits = Math.max((long) normalizedRate.precision() - normalizedRate.scale(), 0L);
 
-        int fractionalDigits = Math.max(
-                normalizedRate.scale(),
-                0
-        );
+        long fractionalDigits = Math.max((long) normalizedRate.scale(), 0L);
 
         return integerDigits > 6 || fractionalDigits > 6;
     }
